+++ WordPress/src/main/java/org/wordpress/android/util/ViewUtils.java
@@ -1,12 +1,17 @@
 package org.wordpress.android.util;
 
+import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.content.res.TypedArray;
+import android.graphics.Outline;
+import android.os.Build;
 import android.support.annotation.AttrRes;
+import android.support.annotation.NonNull;
 import android.support.annotation.StyleRes;
 import android.support.v4.view.ViewCompat;
 import android.view.View;
+import android.view.ViewOutlineProvider;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -53,5 +58,21 @@ public class ViewUtils {
         ColorStateList color = a.getColorStateList(0);
         a.recycle();
         ViewCompat.setBackgroundTintList(button, color);
+    }
+
+    /**
+     * adds an inset circular shadow outline the passed view (Lollipop+ only) - note that
+     * the view should have its elevation set prior to calling this
+     */
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    public static void addCircularShadowOutline(@NonNull View view) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            view.setOutlineProvider(new ViewOutlineProvider() {
+                @Override
+                public void getOutline(View view, Outline outline) {
+                    outline.setOval(0, 0, view.getWidth(), view.getHeight());
+                }
+            });
+        }
     }
 }
