+++ WordPress/src/main/res/layout/media_grid_item.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                xmlns:app="http://schemas.android.com/apk/res-auto"
-                xmlns:tools="http://schemas.android.com/tools"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:padding="@dimen/margin_extra_small">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:padding="@dimen/margin_extra_small">
 
     <org.wordpress.android.widgets.WPNetworkImageView
         android:id="@+id/media_grid_item_image"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:scaleType="centerCrop"/>
+        android:scaleType="centerCrop" />
 
     <RelativeLayout
         android:id="@+id/media_grid_item_file_container"
@@ -34,7 +34,7 @@
                 android:layout_height="24dp"
                 android:layout_gravity="center_horizontal"
                 android:tint="@color/grey_dark"
-                tools:src="@drawable/ic_gridicons_page"/>
+                tools:src="@drawable/ic_gridicons_page" />
 
             <org.wordpress.android.widgets.WPTextView
                 android:id="@+id/media_grid_item_filetype"
@@ -46,7 +46,7 @@
                 android:textColor="@color/grey_dark"
                 android:textSize="@dimen/text_sz_small"
                 android:textStyle="bold"
-                tools:text="PDF"/>
+                tools:text="PDF" />
 
             <org.wordpress.android.widgets.WPTextView
                 android:id="@+id/media_grid_item_name"
@@ -58,7 +58,7 @@
                 android:singleLine="true"
                 android:textColor="@color/grey_dark"
                 android:textSize="@dimen/text_sz_extra_small"
-                tools:text="filename"/>
+                tools:text="filename" />
         </LinearLayout>
 
     </RelativeLayout>
@@ -95,7 +95,7 @@
                     android:layout_marginRight="@dimen/margin_large"
                     android:background="@drawable/photo_picker_circle_normal"
                     android:padding="@dimen/margin_small"
-                    app:srcCompat="@drawable/media_retry_image"/>
+                    app:srcCompat="@drawable/media_retry_image" />
 
                 <ImageView
                     android:id="@+id/image_trash"
@@ -105,7 +105,7 @@
                     android:layout_marginRight="@dimen/margin_large"
                     android:background="@drawable/photo_picker_circle_normal"
                     android:padding="@dimen/margin_small"
-                    app:srcCompat="@drawable/ic_trash_white_24dp"/>
+                    app:srcCompat="@drawable/ic_trash_white_24dp" />
             </LinearLayout>
 
             <ProgressBar
@@ -115,7 +115,7 @@
                 android:layout_gravity="center_horizontal"
                 android:layout_marginBottom="@dimen/margin_small"
                 android:visibility="gone"
-                tools:visibility="visible"/>
+                tools:visibility="visible" />
 
             <org.wordpress.android.widgets.WPTextView
                 android:id="@+id/media_grid_item_upload_state"
@@ -125,7 +125,7 @@
                 android:textColor="@color/white"
                 android:textSize="@dimen/text_sz_medium"
                 android:textStyle="bold"
-                tools:text="@string/media_upload_state_failed"/>
+                tools:text="@string/media_upload_state_failed" />
 
         </LinearLayout>
 
@@ -143,10 +143,11 @@
             android:layout_width="@dimen/media_selection_dot"
             android:layout_height="@dimen/media_selection_dot"
             android:background="@drawable/photo_picker_circle_selector"
+            android:elevation="6dp"
             android:gravity="center"
             android:textColor="@color/white"
-            android:textSize="@dimen/text_sz_medium"
-            tools:text="5"/>
+            android:textSize="@dimen/text_sz_small"
+            tools:text="5" />
     </FrameLayout>
 
     <FrameLayout
@@ -162,7 +163,7 @@
             android:layout_width="@dimen/photo_picker_preview_icon"
             android:layout_height="@dimen/photo_picker_preview_icon"
             android:padding="@dimen/margin_extra_small"
-            app:srcCompat="@drawable/ic_play_video"/>
+            app:srcCompat="@drawable/ic_play_video" />
 
     </FrameLayout>
